添加了数据的词汇云图，更加直观；添加了Restful API的documentation界面；优化了前端。

+++ src/main/java/cse/java2/project/model/Answers.java
@@ -7,7 +7,7 @@ import javax.persistence.Id;
 public class Answers {
 
   @Id
-  public Long answer_id;
-  public Long question_id;
+  public Integer answer_id;
+  public Integer question_id;
   public String body;
 }
